requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
discord==2.0.0
discord.py==2.0.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.2
pylint==2.3.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
websockets==6.0
wrapt==1.11.1
yarl==1.3.0
bs4==0.0.1
selenium==4.4.3
mysql-connector-python==8.0.30